Start using vobject instead of icalendar
Big changes are here. We created new Event class for taking care of
events and ICal for searching events.

Vobject is much nicer than icalendar to use, seemingly with proper
handling of timezones. This needs to be verified later, but date
handling will be refactored (again).

Proper examples will follow